Update README.md
Basically, 2048 presents with with a 4×4 grid. When you start the game, there will be two “tiles” on the grid, each displaying the number 1. You hit the arrow keys on your keyboard to move the tiles around — and also to generate new tiles, which will also be valued at 1. When two tiles having consecutive fibonacci numbers, they combine to give you one greater tile that displays their sum. The more you do this, obviously, the higher the tiles get and the more crowded the board becomes. Your objective is to reach 2584 before the board fills up.